Fix bug when there is no security policy filtered in garbage collector
Process exits reporting panic: runtime error: index out of range [0] with length 0.
The groups are not deleted when deleting security policy.
